pyproject.toml
[tool.poetry]
name = "fast-api-auth"
version = "0.1.0"
description = ""
authors = ["principle105"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.0"
python-dotenv = "^0.19.2"
uvicorn = "^0.15.0"
PyJWT = "^2.3.0"
Jinja2 = "^3.0.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
black = "^21.11b1"
motor = "^2.5.1"
dnspython = "^2.1.0"
certifi = "^2021.10.8"
twilio = "^7.3.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
